create base of diesel

+++ src-tauri/src/data_model/media.rs
@@ -10,9 +10,9 @@ use itertools::Itertools;
 
 #[derive(Debug, Clone, Default, Eq, serde::Serialize)]
 pub struct Media {
-    pub id: i64,
+    pub id: i32,
     pub name: String,
-    pub year: Option<u32>,
+    pub year: Option<i32>,
     pub watched: bool,
     pub my_ranking: u8,
     pub watch_list: bool,
@@ -165,7 +165,7 @@ impl Media {
             .to_string()
     }
 
-    fn detect_year(input: &str) -> Option<u32> {
+    fn detect_year(input: &str) -> Option<i32> {
         let re = Regex::new(r"(19|20)\d{2}").unwrap();
         let mut last = None;
 
@@ -182,7 +182,7 @@ impl Media {
 
             if !is_before_digit
                 && !is_after_digit
-                && let Ok(year) = s.parse::<u32>()
+                && let Ok(year) = s.parse::<i32>()
                 && (1900..=2099).contains(&year)
             {
                 last = Some(year);
